feat: discount usages & notifications, change: double->bigdecimal

+++ src/main/java/com/example/spring_boot_project_api/dto/request/vehicle/VehicleRequestDTO.java
@@ -47,7 +47,7 @@ public class VehicleRequestDTO {
   private BigDecimal pricePerDay;
 
   @NotNull
-  private Double mileAge;
+  private BigDecimal mileAge;
 
   @NotBlank
   private String description;
